fastEDM

The fastEDM package implements a series of tools that can be used for empirical dynamic modeling. The core algorithm is written in C++ to achieve a reasonable execution speed.

References

Jinjing Li, Michael J. Zyphur, George Sugihara, Patrick J. Laub (2021), Beyond Linearity, Stability, and Equilibrium: The edm Package for Empirical Dynamic Modeling and Convergent Cross Mapping in Stata, Stata Journal, 21(1), pp. 220-258

See also

Author

Patrick Laub <patrick.laub@gmail.com>

Examples

library(fastEDM)
library(ggplot2)
#> Error in library(ggplot2): there is no package called ‘ggplot2’
library(readr)
#> Error in library(readr): there is no package called ‘readr’

data <- url("https://raw.githubusercontent.com/EDM-Developers/EDM/master/test/chicago.csv")
chicago <- read_csv(data, col_types = cols(crime = col_double()))
#> Error in read_csv(data, col_types = cols(crime = col_double())): could not find function "read_csv"

libs <- c(seq(10, 200, 5), seq(210, 1000, 10), seq(1020, 2000, 20),
          seq(2050, 4350, 50), 4365)

res1 <- edm(chicago["t"], chicago["temp"], chicago["crime"],
                       E=7, library=libs, numReps=4, verbosity=1, numThreads=4)
#> Error in edm(chicago["t"], chicago["temp"], chicago["crime"], E = 7, library = libs,     numReps = 4, verbosity = 1, numThreads = 4): object 'chicago' not found

res2 <- edm(chicago["t"], chicago["crime"], chicago["temp"],
                       E=7, library=libs, numReps=4, verbosity=1, numThreads=4)
#> Error in edm(chicago["t"], chicago["crime"], chicago["temp"], E = 7, library = libs,     numReps = 4, verbosity = 1, numThreads = 4): object 'chicago' not found

averaged1 <- aggregate(res1$summary[, c("mae", "rho")], list(res1$summary$library), mean)
#> Error in aggregate(res1$summary[, c("mae", "rho")], list(res1$summary$library),     mean): object 'res1' not found
averaged2 <- aggregate(res2$summary[, c("mae", "rho")], list(res2$summary$library), mean)
#> Error in aggregate(res2$summary[, c("mae", "rho")], list(res2$summary$library),     mean): object 'res2' not found
colnames(averaged1)[[1]] <- "library"
#> Error in colnames(averaged1)[[1]] <- "library": object 'averaged1' not found
colnames(averaged2)[[1]] <- "library"
#> Error in colnames(averaged2)[[1]] <- "library": object 'averaged2' not found

p <- ggplot() + 
  geom_line(data = averaged1, aes(x = library, y = rho), color = 2) +
  geom_line(data = averaged2, aes(x = library, y = rho), color = 3) +
  xlab("Library") +
  ylab("Rho")
#> Error in ggplot(): could not find function "ggplot"

p <- p + 
  geom_point(data = res1$summary, aes(x = library, y = rho), alpha = 0.05, color = 2) +
  geom_point(data = res2$summary, aes(x = library, y = rho), alpha = 0.05, color = 3) +
  geom_smooth(method = "loess", alpha = 0.9)
#> Error in eval(expr, envir, enclos): object 'p' not found

print(p)
#> Error in print(p): object 'p' not found